Move Scripter over to LanguageService
