+++ src/Microsoft.SqlTools.LanguageService/Scripting/PeekDefinitionTempFolder.cs
@@ -1,0 +1,73 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+#nullable disable
+
+using System;
+using System.IO;
+
+namespace Microsoft.SqlTools.LanguageService.Scripting
+{
+    /// <summary>
+    /// Manages the per-process temporary folder used to store scripts generated by Peek Definition.
+    /// </summary>
+    internal static class PeekDefinitionTempFolder
+    {
+        private static readonly object TempFolderLock = new object();
+
+        /// <summary>
+        /// The root path of the temporary folder used to store generated scripts. Updated to the
+        /// process-specific folder the first time <see cref="GetTempFolder"/> is called.
+        /// </summary>
+        internal static string TempFolderPath = Path.GetTempPath() + "mssql_definition";
+
+        internal static bool TempFolderCreated = false;
+
+        /// <summary>
+        /// Returns the temporary folder used to store generated scripts, creating a private
+        /// process-specific folder once so concurrent peek-definition requests share the same root.
+        /// </summary>
+        internal static string GetTempFolder()
+        {
+            lock (TempFolderLock)
+            {
+                string tempPath;
+                if (!TempFolderCreated)
+                {
+                    try
+                    {
+                        // create a private temp folder once per process so concurrent peek-definition requests share the same root
+                        string tempFolder = string.Format("{0}_{1}", TempFolderPath, Guid.NewGuid().ToString("N"));
+                        DirectoryInfo tempScriptDirectory = Directory.CreateDirectory(tempFolder);
+                        TempFolderPath = tempScriptDirectory.FullName;
+                        tempPath = tempScriptDirectory.FullName;
+                        TempFolderCreated = true;
+                    }
+                    catch (Exception)
+                    {
+                        // swallow exception and use temp folder to store scripts
+                        tempPath = Path.GetTempPath();
+                    }
+                }
+                else
+                {
+                    try
+                    {
+                        // use tempDirectory name created previously
+                        DirectoryInfo tempScriptDirectory = Directory.CreateDirectory(TempFolderPath);
+                        tempPath = tempScriptDirectory.FullName;
+                    }
+                    catch (Exception)
+                    {
+                        // swallow exception and use temp folder to store scripts
+                        tempPath = Path.GetTempPath();
+                    }
+                }
+
+                return tempPath;
+            }
+        }
+    }
+}
