+++ test/Microsoft.SqlTools.ServiceLayer.UnitTests/Utility/SrTests.cs
@@ -35,22 +35,9 @@ namespace Microsoft.SqlTools.ServiceLayer.UnitTests.Utility
             Assert.NotNull(ServiceLayerSr.ConnectionParamsValidateNullSqlAuth(""));
             Assert.NotNull(ServiceLayerSr.ConnectionServiceConnectErrorNullParams);
             Assert.NotNull(ServiceLayerSr.ConnectionServiceConnectionCanceled);
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringMissingAttestationUrlWithAttestationProtocol);
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidAttestationProtocolNoneWithUrl);
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringMissingAttestationProtocolWithSecureEnclaves);
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidEnclaveAttestationProtocol(""));
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidColumnEncryptionSetting(""));
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidSecureEnclaves(""));
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidEncryptOption(""));
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidAuthType(""));
-            Assert.NotNull(ServiceLayerSr.ConnectionServiceConnStringInvalidIntent(""));
             Assert.NotNull(ServiceLayerSr.ConnectionServiceDbErrorDefaultNotConnected(""));
             Assert.NotNull(ServiceLayerSr.ConnectionServiceListDbErrorNotConnected(""));
             Assert.NotNull(ServiceLayerSr.ConnectionServiceListDbErrorNullOwnerUri);
-            Assert.NotNull(ServiceLayerSr.PeekDefinitionAzureError(""));
-            Assert.NotNull(ServiceLayerSr.PeekDefinitionDatabaseError);
-            Assert.NotNull(ServiceLayerSr.PeekDefinitionError(""));
-            Assert.NotNull(ServiceLayerSr.PeekDefinitionNoResultsError);
             Assert.NotNull(ServiceLayerSr.PeekDefinitionNotConnectedError);
             Assert.NotNull(ServiceLayerSr.PeekDefinitionTimedoutError);
             Assert.NotNull(ServiceLayerSr.QueryServiceAffectedOneRow);
